+++ src/index.ts
@@ -9,11 +9,15 @@ export {
 } from './lib/provable/foreign-field.js';
 export {
   createForeignCurve,
+  createForeignCurveV2,
   ForeignCurve,
+  ForeignCurveV2,
 } from './lib/provable/crypto/foreign-curve.js';
 export {
   createEcdsa,
+  createEcdsaV2,
   EcdsaSignature,
+  EcdsaSignatureV2,
 } from './lib/provable/crypto/foreign-ecdsa.js';
 export {
   Poseidon,